added book listing html

+++ app/models/token.rb
@@ -15,27 +15,53 @@ class Token < ActiveRecord::Base
 		# redeemer.credit_transactions.create(:giver_id => self.owner,:message => AppConfiguration::TOKEN_REDEMPTION,:credit => self.credit,:token_id=>self.id)
 		self.is_redeemed = true
 		self.save
-		Token.generate_token(redeemer)
-		if self.receiver_id.present? && self.book_id.present?
+		Token.generate_book_coin(redeemer)
+		if self.receiver_id.present? && self.book_id.present? && owner.present?
 			book_user = BooksUser.find_by_book_id_and_user_id(self.book_id,self.receiver_id)
 			book_user.locations.destroy_all
 			book_user.is_provided = false
 			book_user.save
+			new_book_user = BooksUser.find_by_book_id_and_user_id(self.book_id,owner)
+			if new_book_user.present?
+				new_book_user.count += 1
+			else
+				new_book_user = BooksUser.create(:book_id => self.book_id,:user_id => owner,:is_provided => false)
+			end
+
 			# self.giver.profile.credit -= self.credit
 			# self.giver.profile.save
 			# self.giver.debit_transactions.create(:receiver_id => redeemer.id,:message => AppConfiguration::TOKEN_REDEMPTION,:credit => self.credit,:token_id=>self.id)
 		end
 	end
 
-	def check_validity
-		return !self.is_redeemed && self.valid_til >= Date.today
+	def check_validity(options = {})
+		if options[:receiver_id].present?
+			return self.receiver_id == options[:receiver_id] && !self.is_redeemed && self.valid_til >= Date.today
+		else
+			return !self.is_redeemed && self.valid_til >= Date.today
+		end
+		
 	end
 
-	def self.generate_token(user)
+	def self.generate_book_coin(user)
+		token = user.tokens.new(:credit => AppConfiguration::TOKEN_GENERATION_CREDIT)
+	    token.save
+	end
+
+	def generate_token
 	    raw_string = SecureRandom.random_number( 2**80 ).to_s( 20 ).reverse
 	    long_code = raw_string.tr( '0123456789abcdefghij', '234679QWERTYUPADFGHX' )
 	    short_code = long_code[0..3] + '-' + long_code[4..7] + '-' + long_code[8..11]
-	    token = user.tokens.new(:redeem_code => short_code,:credit => AppConfiguration::TOKEN_GENERATION_CREDIT)
-	    token.save
+	    return short_code
+	end
+
+	def available_unlocked
+		!self.is_redeemed? && !self.book_id.present? && !self.receiver_id.present?
+	end
+	def locked
+		!self.is_redeemed? && self.book_id.present? && self.receiver_id.present?
+	end
+	def redeemed
+		self.is_redeemed?
 	end
 end
